pyproject.toml
[project]
name = "code_review_ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Liubomyr",email = "liubomyr.oleksiuk@ucu.edu.ua"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "httpx (>=0.27.2,<0.28.0)",
    "pydantic (>=2.10.5,<3.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pytest-httpx (>=0.30.0,<0.36.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "google-generativeai (>=0.8.3,<0.9.0)",
    "pytest-asyncio (>=0.25.2,<0.26.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "setuptools (>=75.8.0,<76.0.0)",
    "redis (>=5.2.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
